FIX: Port over anatomical derivative workflow
Hack to ensure template resolution does not default to 1 in cases where template is "res-less".

Currently, TemplateFlow does not have res(olution) identifiers for:
- UNCInfant (all cohorts)
- Fischer344

In many places, templates are assumed to have the `res` entity.
An easy fix would be to add the `res` entity to these templates, but now may be time to implement "on-the-fly" resolution...