FIX: Name iternode/joinnode something unique
There is a bug deep within nipype workflow connection / join expansions that either:
- mistakenly uses node.name rather than node._hierarchy to fetch the relevant node or
- reuses a previous JoinNode's `_next_slot_index`

In either case, the end result is some funky indexed connection (in this case it was looking for the fields `bold_fsLRJ3` and `bold_fsLRJ4`.

I spent some time in the rabbit hole and came out empty-handed, but the easiest solution seems to be just renaming them to something unique. Pain.